+++ sdk/storage/Azure.Storage.DataMovement/CHANGELOG.md
@@ -10,6 +10,7 @@
 - Changed `TransferCheckpointStoreOptions.CheckpointerPath` to internal.
 
 ### Bugs Fixed
+- Fixed bug where adding multiple transfers in parallel could cause a collision (`InvalidOperationException`) in the data transfers stored within the `TransferManager`.
 
 ### Other Changes